fix(core): bundle the genome assemblies into the JavaScript library
Built-in genome assemblies were loaded from https://genomespy.app/. This is not a good idea
because it makes the library dependent on the availability of the server. Instead, the genome
assemblies are now bundled into the JavaScript library. Custom genomes are still supported
but the way they are specified has changed a bit.